[TEST-ONLY] Migrate IncrementalCompilationTests to Swift Testing
Migrate the incremental compilation tests from XCTest to Swift Testing,
splitting the monolithic IncrementalCompilationTests.swift into:

- IncrementalBasicTests: core incremental build tests
- IncrementalExplicitBuildTests: explicit module / caching tests
- IncrementalInputModificationTests: input add/remove/reorder tests

Extract IncrementalTestHarness into its own file and introduce
TestBuildConfig to parameterize explicit-module tests over implicit,
explicit, caching, and prefix-mapped caching configurations.
